pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "curriculum-flow-env"
version = "1.0.0"
description = "Multi-agent RL environment for adaptive curriculum optimization — OpenEnv compatible"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{name = "krithickvivek"}]
keywords = ["reinforcement-learning", "education", "openenv", "gymnasium", "adaptive-learning"]

dependencies = [
    "openenv-core>=0.2.0",
    "gymnasium>=0.29.0",
    "numpy>=1.24.0",
    "networkx>=3.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
server = [
    "python-multipart>=0.0.6",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "huggingface_hub>=0.20.0",
    "plotly>=5.18.0",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[tool.setuptools.packages.find]
include = ["curriculum_flow_env*", "models*", "client*", "server*"]
